update script name

+++ src/andushu/dataset_readers/math/math2tree.py
@@ -1,4 +1,5 @@
 import json
+import os
 import re
 import warnings
 from typing import List, Dict
@@ -208,20 +209,27 @@ class Math2TreeDatasetReader(DatasetReader):
     def _read_math23k(self, file_path, op_type):
         errors = []
         total = 0
-        with jsonlines.open(file_path + f'.{op_type}.jsonl', mode="w") as writer:
-            with open(file_path, encoding="utf-8") as f:
-                for item in json.load(f):
+        filename = file_path + f'.{op_type}.jsonl'
+        if os.path.isfile(filename):
+            with jsonlines.open(filename) as reader:
+                for item in reader:
                     total += 1
-                    item = Processor.process_math23k(self.ast_parser, item,
-                                                     use_chinese_segmentation=self._chinese_segmentation)
-                    if item:
-                        if op_type == "disallow_pow" and "Pow" in item['equation']:
-                            continue
+                    yield item
+        else:
+            with jsonlines.open(file_path + f'.{op_type}.jsonl', mode="w") as writer:
+                with open(file_path, encoding="utf-8") as f:
+                    for item in json.load(f):
+                        total += 1
+                        item = Processor.process_math23k(self.ast_parser, item,
+                                                         use_chinese_segmentation=self._chinese_segmentation)
+                        if item:
+                            if op_type == "disallow_pow" and "Pow" in item['equation']:
+                                continue
 
-                        writer.write(item)
-                        yield item
-                    else:
-                        errors.append(item)
+                            writer.write(item)
+                            yield item
+                        else:
+                            errors.append(item)
         logger.info(f"Total instances: {total} \n"
                     f"Error instances: {len(errors)} \n"
                     f"Loaded instances: {total - len(errors)}")
@@ -229,28 +237,36 @@ class Math2TreeDatasetReader(DatasetReader):
     def _read_mathqa(self, file_path, op_type):
         errors = []
         total = 0
-        with jsonlines.open(file_path + f'.{op_type}.jsonl', mode="w") as writer:
-            with open(file_path, encoding="utf-8") as f:
-                for item in json.load(f):
-                    total += 1
-                    try:
-                        item = Processor.process_mathqa(self.ast_parser, item, filtered_ops[op_type])
-                        status = 'ok' if item is not None else 'err'
-                    except SyntaxError:
-                        status = 'err'
-                    except ZeroDivisionError:
-                        status = 'err'
-                    except ValueError:
-                        status = 'err'
-                    except Exception as e:
-                        status = 'err'
 
-                    if status == 'ok':
-                        item['process_type'] = 'mathqa'
-                        writer.write(item)
-                        yield item
-                    else:
-                        errors.append(item)
+        filename = file_path + f'.{op_type}.jsonl'
+        if os.path.isfile(filename):
+            with jsonlines.open(filename) as reader:
+                for item in reader:
+                    total += 1
+                    yield item
+        else:
+            with jsonlines.open(filename, mode="w") as writer:
+                with open(file_path, encoding="utf-8") as f:
+                    for item in json.load(f):
+                        total += 1
+                        try:
+                            item = Processor.process_mathqa(self.ast_parser, item, filtered_ops[op_type])
+                            status = 'ok' if item is not None else 'err'
+                        except SyntaxError:
+                            status = 'err'
+                        except ZeroDivisionError:
+                            status = 'err'
+                        except ValueError:
+                            status = 'err'
+                        except Exception as e:
+                            status = 'err'
+
+                        if status == 'ok':
+                            item['process_type'] = 'mathqa'
+                            writer.write(item)
+                            yield item
+                        else:
+                            errors.append(item)
         logger.info(f"Total instances: {total} \n"
                     f"Error instances: {len(errors)} \n"
                     f"Loaded instances: {total - len(errors)}")
